refactor: add `inference_utils` module
