+++ test_unstructured/partition/pdf_image/test_ocr.py
@@ -4,13 +4,8 @@ import numpy as np
 import pandas as pd
 import pytest
 import unstructured_pytesseract
-from pdf2image.exceptions import PDFPageCountError
 from PIL import Image, UnidentifiedImageError
-from unstructured_inference.inference.elements import EmbeddedTextRegion, TextRegion
-from unstructured_inference.inference.layout import DocumentLayout
-from unstructured_inference.inference.layoutelement import (
-    LayoutElement,
-)
+from pdf2image.exceptions import PDFPageCountError
 
 from unstructured.documents.elements import ElementType
 from unstructured.partition.pdf_image import ocr
@@ -18,14 +13,15 @@ from unstructured.partition.pdf_image.ocr import pad_element_bboxes
 from unstructured.partition.utils.constants import (
     Source,
 )
-from unstructured.partition.utils.ocr_models.ocr_interface import (
-    get_elements_from_ocr_regions,
-    merge_text_regions,
-)
 from unstructured.partition.utils.ocr_models.paddle_ocr import OCRAgentPaddle
 from unstructured.partition.utils.ocr_models.tesseract_ocr import (
     OCRAgentTesseract,
     zoom_image,
+)
+from unstructured_inference.inference.elements import EmbeddedTextRegion, TextRegion
+from unstructured_inference.inference.layout import DocumentLayout
+from unstructured_inference.inference.layoutelement import (
+    LayoutElement,
 )
 
 
@@ -229,35 +225,6 @@ def test_aggregate_ocr_text_by_block():
 
     text = ocr.aggregate_ocr_text_by_block(ocr_layout, region, 0.5)
     assert text == expected
-
-
-def test_merge_text_regions(mock_embedded_text_regions):
-    expected = TextRegion.from_coords(
-        x1=437.83888888888885,
-        y1=317.319341111111,
-        x2=1256.334784222222,
-        y2=406.9837855555556,
-        text="LayoutParser: A Unified Toolkit for Deep Learning Based Document Image",
-    )
-
-    merged_text_region = merge_text_regions(mock_embedded_text_regions)
-    assert merged_text_region == expected
-
-
-def test_get_elements_from_ocr_regions(mock_embedded_text_regions):
-    expected = [
-        LayoutElement.from_coords(
-            x1=437.83888888888885,
-            y1=317.319341111111,
-            x2=1256.334784222222,
-            y2=406.9837855555556,
-            text="LayoutParser: A Unified Toolkit for Deep Learning Based Document Image",
-            type=ElementType.UNCATEGORIZED_TEXT,
-        ),
-    ]
-
-    elements = get_elements_from_ocr_regions(mock_embedded_text_regions)
-    assert elements == expected
 
 
 @pytest.mark.parametrize("zoom", [1, 0.1, 5, -1, 0])
