+++ unstructured/partition/utils/ocr_models/paddle_ocr.py
@@ -1,19 +1,20 @@
-from typing import List
+from typing import List, TYPE_CHECKING
 
 import numpy as np
 from PIL import Image as PILImage
-from unstructured_inference.inference.elements import TextRegion
-from unstructured_inference.inference.layoutelement import (
-    LayoutElement,
-)
 
 from unstructured.documents.elements import ElementType
 from unstructured.logger import logger
+from unstructured.partition.pdf_image.inference_utils import build_layout_element
 from unstructured.partition.utils.constants import (
     DEFAULT_PADDLE_LANG,
     Source,
 )
 from unstructured.partition.utils.ocr_models.ocr_interface import OCRAgent
+
+if TYPE_CHECKING:
+    from unstructured_inference.inference.elements import TextRegion
+    from unstructured_inference.inference.layoutelement import LayoutElement
 
 
 class OCRAgentPaddle(OCRAgent):
@@ -63,7 +64,7 @@ class OCRAgentPaddle(OCRAgent):
 
     def get_layout_from_image(
         self, image: PILImage, ocr_languages: str = "eng"
-    ) -> List[TextRegion]:
+    ) -> List["TextRegion"]:
         """Get the OCR regions from image as a list of text regions with paddle."""
 
         logger.info("Processing entire page OCR with paddle...")
@@ -78,7 +79,7 @@ class OCRAgentPaddle(OCRAgent):
 
     def get_layout_elements_from_image(
         self, image: PILImage, ocr_languages: str = "eng"
-    ) -> List[LayoutElement]:
+    ) -> List["LayoutElement"]:
         ocr_regions = self.get_layout_from_image(
             image,
             ocr_languages=ocr_languages,
@@ -88,13 +89,11 @@ class OCRAgentPaddle(OCRAgent):
         # terms of grouping because we get ocr_text from `ocr_layout, so the first two grouping
         # and merging steps are not necessary.
         return [
-            LayoutElement(
-                bbox=r.bbox, text=r.text, source=r.source, type=ElementType.UNCATEGORIZED_TEXT
-            )
+            build_layout_element(bbox=r.bbox, text=r.text, source=r.source, element_type=ElementType.UNCATEGORIZED_TEXT)
             for r in ocr_regions
         ]
 
-    def parse_data(self, ocr_data: list) -> List[TextRegion]:
+    def parse_data(self, ocr_data: list) -> List["TextRegion"]:
         """
         Parse the OCR result data to extract a list of TextRegion objects from
         paddle.
@@ -114,6 +113,7 @@ class OCRAgentPaddle(OCRAgent):
         - An empty string or a None value for the 'text' key in the input
           dictionary will result in its associated bounding box being ignored.
         """
+
         text_regions = []
         for idx in range(len(ocr_data)):
             res = ocr_data[idx]
